Fix loading of model library in non-electron env
