+++ src/components/common/TreeExplorerTreeNode.vue
@@ -10,27 +10,25 @@
     ]"
     ref="container"
   >
-    <div class="flex flex-col w-full">
-      <div class="node-content">
-        <span class="node-label">
-          <slot name="before-label" :node="props.node"></slot>
-          <EditableText
-            :modelValue="node.label"
-            :isEditing="isEditing"
-            @edit="handleRename"
-          />
-          <slot name="after-label" :node="props.node"></slot>
-        </span>
-        <Badge
-          v-if="showNodeBadgeText"
-          :value="nodeBadgeText"
-          severity="secondary"
-          class="leaf-count-badge"
+    <div class="node-content">
+      <span class="node-label">
+        <slot name="before-label" :node="props.node"></slot>
+        <EditableText
+          :modelValue="node.label"
+          :isEditing="isEditing"
+          @edit="handleRename"
         />
-      </div>
-      <div class="node-actions">
-        <slot name="actions" :node="props.node"></slot>
-      </div>
+        <slot name="after-label" :node="props.node"></slot>
+      </span>
+      <Badge
+        v-if="showNodeBadgeText"
+        :value="nodeBadgeText"
+        severity="secondary"
+        class="leaf-count-badge"
+      />
+    </div>
+    <div class="node-actions">
+      <slot name="actions" :node="props.node"></slot>
     </div>
   </div>
 </template>
