+++ src/components/sidebar/tabs/ModelLibrarySidebarTab.vue
@@ -28,15 +28,7 @@
       />
     </template>
     <template #body>
-      <div class="mx-6 mb-4" v-if="downloads.length > 0">
-        <div class="text-lg my-4">
-          {{ t('electronFileDownload.inProgress') }}
-        </div>
-
-        <template v-for="download in downloads" :key="download.url">
-          <DownloadItem :download="download" />
-        </template>
-      </div>
+      <ElectronDownloadItems v-if="isElectron()" />
 
       <TreeExplorer
         class="model-lib-tree-explorer py-0"
@@ -58,7 +50,7 @@ import SearchBox from '@/components/common/SearchBox.vue'
 import TreeExplorer from '@/components/common/TreeExplorer.vue'
 import SidebarTabTemplate from '@/components/sidebar/tabs/SidebarTabTemplate.vue'
 import ModelTreeLeaf from '@/components/sidebar/tabs/modelLibrary/ModelTreeLeaf.vue'
-import DownloadItem from '@/components/sidebar/tabs/modelLibrary/DownloadItem.vue'
+import ElectronDownloadItems from '@/components/sidebar/tabs/modelLibrary/ElectronDownloadItems.vue'
 import {
   ComfyModelDef,
   ModelFolder,
@@ -76,9 +68,7 @@ import { computed, ref, watch, toRef, onMounted, nextTick } from 'vue'
 import type { TreeNode } from 'primevue/treenode'
 import { app } from '@/scripts/app'
 import { buildTree } from '@/utils/treeUtil'
-import { useI18n } from 'vue-i18n'
-import { useElectronDownloadStore } from '@/stores/electronDownloadStore'
-import { storeToRefs } from 'pinia'
+import { isElectron } from '@/utils/envUtil'
 
 const modelStore = useModelStore()
 const modelToNodeStore = useModelToNodeStore()
@@ -86,9 +76,6 @@ const settingStore = useSettingStore()
 const searchQuery = ref<string>('')
 const expandedKeys = ref<Record<string, boolean>>({})
 const { expandNode, toggleNodeOnEvent } = useTreeExpansion(expandedKeys)
-const { t } = useI18n()
-const electronDownloadStore = useElectronDownloadStore()
-const { downloads } = storeToRefs(electronDownloadStore)
 
 const filteredModels = ref<ComfyModelDef[]>([])
 const handleSearch = async (query: string) => {
